controladores y modelos entidades simples. Modelos event y community

+++ app/controllers/province.controller.js
@@ -1,60 +1,66 @@
-const db = require("../models");
-const Province = db.province;
-const Op = db.Sequelize.Op;
+const { Sequelize, Op } = require("sequelize");
+const sequelize = require("../database/database.js");
+const Province = sequelize.models.Province;
 
 // Create and Save a new ciudad
 exports.create = (req, res) => {
-  // Validate request
-  if (!req.body.nombre) {
+  if (!req.body.name) {
     res.status(400).send({
       message: "Content can not be empty!"
     });
     return;
   }
-
-  // Create a Tutorial
   const provincia = {
-    nombre: req.body.nombre
+    name: req.body.name
   };
-
-  // Save Tutorial in the database
   Province.create(provincia)
     .then(data => {
-      res.send(data);
+      res.status(201).send(data);
     })
     .catch(err => {
       res.status(500).send({
         message:
-          err.message || "Some error occurred while creating "
+          err.name + ': ' + err.message || "Some error occurred while creating "
       });
     });
 };
 
 // Retrieve all ciudades from the database.
 exports.findAll = (req, res) => {
-  const nombre = req.query.nombre;
-  var condition = nombre ? { nombre: { [Op.like]: `%${nombre}%` } } : null;
+  const name = req.query.name;
+  var condition = name ? { name: { [Op.like]: `%${name}%` } } : null;
 
   Province.findAll({ where: condition })
     .then(data => {
-      res.send(data);
+      if (data) {
+        res.status(200).send(data);
+      }
+      else {
+        res.status(404).send({ message: 'Cannot find' })
+      }
     })
     .catch(err => {
       res.status(500).send({
         message:
-          err.message || "Some error occurred while retrieving "
+          err.name + ': ' + err.message || "Some error occurred while retrieving "
       });
     });
 };
 
 // Find a single ciudad with an id
 exports.findOne = (req, res) => {
+  if (!req.params.id) {
+    res.status(400).send({
+      message: "id can not be empty!"
+    });
+    return
+  }
   const id = req.params.id;
 
   Province.findByPk(id)
     .then(data => {
       if (data) {
-        res.send(data);
+        res.status(200).send(data);
       } else {
         res.status(404).send({
           message: `Cannot find with id=${id}.`
@@ -63,13 +69,19 @@ exports.findOne = (req, res) => {
     })
     .catch(err => {
       res.status(500).send({
-        message: "Error retrieving  with id=" + id
+        message: err.name + ': ' + err.message || "Error retrieving  with id=" + id
       });
     });
 };
 
 // Update by the id in the request
 exports.update = (req, res) => {
+  if (!req.params.id) {
+    res.status(400).send({
+      message: "id can not be empty!"
+    });
+    return
+  }
   const id = req.params.id;
 
   Province.update(req.body, {
@@ -77,24 +89,30 @@ exports.update = (req, res) => {
   })
     .then(num => {
       if (num == 1) {
-        res.send({
+        res.status(200).send({
           message: "provincia was updated successfully."
         });
       } else {
-        res.send({
+        res.status(404).send({
           message: `Cannot update provincia with id=${id}. Maybe provincia was not found or req.body is empty!`
         });
       }
     })
     .catch(err => {
       res.status(500).send({
-        message: "Error updating provincia with id=" + id
+        message: err.name + ': ' + err.message || "Error updating provincia with id=" + id
       });
     });
 };
 
 // Delete a Tutorial with the specified id in the request
 exports.delete = (req, res) => {
+  if (!req.params.id) {
+    res.status(400).send({
+      message: "id can not be empty!"
+    });
+    return
+  }
   const id = req.params.id;
 
   Province.destroy({
@@ -102,18 +120,18 @@ exports.delete = (req, res) => {
   })
     .then(num => {
       if (num == 1) {
-        res.send({
-          message: "provincia was deleted successfully!"
+        res.status(200).send({
+          message: "The province was deleted successfully!"
         });
       } else {
-        res.send({
+        res.status(404).send({
           message: `Cannot delete provincia with id=${id}. Maybe provincia was not found!`
         });
       }
     })
     .catch(err => {
       res.status(500).send({
-        message: "Could not delete provincia with id=" + id
+        message: err.name + ': ' + err.message || "Could not delete provincia with id=" + id
       });
     });
 };
